requirements.txt
discord.py>=2.3.2
openai==0.28.1
python-dotenv>=1.0.0